requirements.txt
numpy==1.26.4
pandas==2.2.2
plip==2.3.0
beautifulsoup4==4.12.3
requests==2.32.2
openai==1.30.2
rdkit==2023.9.6
pubchempy==1.0.4
vina==1.2.5
biopython==1.83
rich==13.7.1
openbabel==3.1.1
pymol-open-source==3.0.0
